Rename aws to aws_lambda, and improve README
