+++ README.md
@@ -1,6 +1,24 @@
 # `filewatch-trigger`
 
-Experimental File Watcher program to trigger AWS Lambda.
+Experimental File Watcher program to trigger any of the below actions:
+
+- AWS Lambda function
+- Shell template command
+
+## Docker Usage
+
+Visit <https://hub.docker.com/r/guangie88/filewatch-trigger/> to check out the
+pre-built Docker images to immediately use the application.
+
+As an example:
+
+```bash
+docker run --rm -it guangie88/filewatch-trigger:python-3.7 -h
+```
+
+This should present the application help screen for more usage infomation.
+
+### AWS Lambda function
 
 Triggers AWS Lambda based on given function name with a known trigger file path,
 and using default AWS credentials based on `boto3`, such as placing `config` and
@@ -14,26 +32,12 @@ Submits payload of the given JSON form in byte format:
 }
 ```
 
-## Usage
-
-### Docker
-
-Visit <https://hub.docker.com/r/guangie88/filewatch-trigger/> to check out
-the pre-built Docker images to immediately use the application.
-
-As an example:
-
-```bash
-docker run --rm -it guangie88/filewatch-trigger:python-3.7 -h
-```
-
-This should present the application help screen for more usage infomation.
-
 For a quick and easy usage example:
 
 ```bash
 docker run --rm -it -v ~/.aws:/root/.aws -v ${PWD}:/app \
     guangie88/filewatch-trigger:python-3.7 \
+    aws-lambda \
     -p . \
     -f "*.jpg,*.png" \
     -n test-trigger
@@ -53,7 +57,27 @@ def lambda_handler(event, context):
     return 'Path is {}'.format(event['path'])
 ```
 
-### Native
+### Shell template command
+
+For a quick and easy usage example:
+
+```bash
+docker run --rm -it -v ~/.aws:/root/.aws -v ${PWD}:/app \
+    guangie88/filewatch-trigger:python-3.7 \
+    cmd \
+    -p . \
+    -f "*.jpg,*.png" \
+    -c "echo \"(File: {0}, Event: {1})\""
+```
+
+The above command runs the application watching the current host directory
+recursively for newly created files ending with `.jpg` and `.png`, and runs
+the `echo` command with the actual path substituted into `{0}` and event name
+into `{1}`.
+
+An example echo output is as follow: `(File: /app/a.png, Event: created)`.
+
+## Running Native
 
 If you prefer running the application natively using Python 3 without Docker,
 you can simply run the following to get the application running like in the
